refactor: move dashboard interfaces to types.ts
